// src/models/schemas/RefreshToken.schema.ts
import { ObjectId } from 'mongodb'

type RefreshTokenType = {
    _id?: ObjectId
    token: string
    user_id: ObjectId
    created_at?: Date
}

export default class RefreshToken {
    _id?: ObjectId
    token: string
    user_id: ObjectId
    created_at: Date

    constructor(refreshToken: RefreshTokenType) {
        this._id = refreshToken._id
        this.token = refreshToken.token
        this.user_id = refreshToken.user_id
        this.created_at = refreshToken.created_at || new Date()
    }
}